Added functions and buttons for removeFish and removeFromOrder.
The removeFish function and button allow removal of a fish object from the inventory,
which in turn removes it from the app, render, and order, updating the order with a
no longer available message. The removeFromOrder function and x element allow a user
to remove the no longer available message from the order. Bind was needed in order to
pass an individual key(fish) from the app component to the inventory and order components.